+++ source/app/service-providers/commands/force-open.ts
@@ -37,7 +37,6 @@ export default class ForceOpen extends ZettlrCommand {
     // Determine if the file should be created, if it can't be found. For this
     // we need both the respective preferences setting and an auto-search
     // command.
-    const autoCreate: boolean = this._app.config.get('zkn.autoCreateLinkedFiles')
     const customDir: string = this._app.config.get('zkn.customDirectory')
 
     const file = this._app.fsal.findExact(linkContents)
@@ -45,12 +44,10 @@ export default class ForceOpen extends ZettlrCommand {
     // Now we have a file (if not, create a new one if the user wishes so)
     if (file !== undefined) {
       await this._app.documents.openFile(windowId, leafId, file.path, newTab)
-    } else if (autoCreate && isDir(customDir)) {
+    } else if (isDir(customDir)) {
       // Call the file-new command on the application, which'll do all
       // necessary steps for us.
       await this._app.commands.run('file-new', { windowId, leafId, name: linkContents, path: customDir })
-    } else if (autoCreate && !isDir(customDir)) {
-      await this._app.commands.run('file-new', { windowId, leafId, name: linkContents })
     }
   }
 }